% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbhsst-package.R
\docType{package}
\name{bbhsst-package}
\alias{bbhsst-package}
\alias{bbh}
\alias{bbhsst}
\title{bbhsst package}
\description{
Serving BBH SST, must end in period.
}
\section{data source}{

Serves up DMR's data for \href{https://dmr-maine.opendata.arcgis.com/datasets/maine::mainedmr-boothbay-harbor-sea-surface-temperatures/about}{BBH SST}
}

