% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{e01_bydate}
\alias{e01_bydate}
\title{Calculate by-date EO1 average, min and max SST}
\usage{
e01_bydate(x = fetch_E01())
}
\arguments{
\item{x}{df, the E01 dataset}
}
\value{
tibble of data
}
\description{
Calculate by-date EO1 average, min and max SST
}
