% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_E01}
\alias{fetch_E01}
\title{Fetch a table of buoy data between specified dates}
\usage{
fetch_E01(
  begin = "2001-07-09",
  end = format(Sys.Date(), "\%Y-\%m-\%d"),
  form = c("hourly", "daily")[1]
)
}
\arguments{
\item{begin}{character, the starting date as YYYY-mm-dd}

\item{end}{character, the end date as YYYY-mm-dd}

\item{form}{character, daily or hourly data}
}
\value{
tibble of data
}
\description{
Fetch a table of buoy data between specified dates
}
