% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_sst}
\alias{fetch_sst}
\title{Accepts name (BBH or E01) and reads data}
\usage{
fetch_sst(name = "bbh", form = c("daily"))
}
\arguments{
\item{name}{character, the name of the dataset}

\item{form}{character, daily or hourly data}
}
\value{
tibble
}
\description{
Accepts name (BBH or E01) and reads data
}
