% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{make_uri}
\alias{make_uri}
\title{Make the URI needed for the specifed date range}
\usage{
make_uri(begin = "2001-07-09", end = "2022-03-30")
}
\arguments{
\item{begin}{character, the starting date as YYYY-mm-dd}

\item{end}{character, the end date as YYYY-mm-dd}
}
\value{
character URI
}
\description{
Make the URI needed for the specifed date range
}
