% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sst}
\alias{plot.sst}
\title{Plots SST data (single series)}
\usage{
\method{plot}{sst}(x, xlim = range(x$COLLECTION_DATE, na.rm = TRUE), ...)
}
\arguments{
\item{x}{sst class tibble}

\item{xlim}{date range limits, defaults to max/min}

\item{...}{other arguments passed to \code{\link[base]{plot}}}
}
\value{
ggplot object
}
\description{
Plots SST data (single series)
}
