% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sst_group}
\alias{plot.sst_group}
\title{#' Plots SST data
#' 
#' @export
#' @param x sst class tibble
#' @param y NULL, ignored, or another tibble of class sst
#' @param type character see \code{\link[base]{plot}}, defaults to 'l'
#' @param ... other arguments passed to \code{\link[base]{plot}} 
#' @return NULL}
\usage{
\method{plot}{sst_group}(x, ...)
}
\arguments{
\item{x}{sst class tibble}

\item{...}{other arguments passed to \code{\link[base]{plot}}}
}
\value{
ggplot object
}
\description{
plot.sst <- function(x, y = NULL, type = 'l', ...){
  plot(x$COLLECTION_DATE, x$SEA_SURFACE_TEMP_AVG_C, type = type, ...)
}
Plots SST data (multiple series)
}
