% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rbind.sst}
\alias{rbind.sst}
\title{Row-binds SST data}
\usage{
\method{rbind}{sst}(...)
}
\arguments{
\item{...}{one or more sst class tibble passed to \code{\link[base]{rbind}}}
}
\value{
a tibble of class sst_group
}
\description{
Row-binds SST data
}
